opt: move stats printing function to Statistics
Previously, the logic to print Statistics was contained in a
private method inside memo/expr_view.go, making it difficult to
debug issues with the stats code. This commit moves the logic into
a String method on Statistics.

Release note: None
