+++ pkg/sql/opt/memo/statistics_builder_test.go
@@ -17,15 +17,12 @@ package memo
 import (
 	"testing"
 
-	"strings"
-
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/constraint"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/props"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/util"
-	"github.com/cockroachdb/cockroach/pkg/util/treeprinter"
 )
 
 // Most of the functionality in statistics.go is tested by the data-driven
@@ -118,63 +115,63 @@ func TestGetStatsFromConstraint(t *testing.T) {
 	cs1 := constraint.SingleConstraint(&c1)
 	statsFunc123(
 		cs1,
-		"stats: [rows=140000000, distinct(1)=7]",
+		"[rows=140000000, distinct(1)=7]",
 		7.0/500,
 	)
 
 	cs2 := constraint.SingleConstraint(&c2)
 	statsFunc123(
 		cs2,
-		"stats: [rows=3333333333]",
+		"[rows=3333333333]",
 		1.0/3,
 	)
 
 	cs3 := constraint.SingleConstraint(&c3)
 	statsFunc123(
 		cs3,
-		"stats: [rows=20000000, distinct(3)=1]",
+		"[rows=20000000, distinct(3)=1]",
 		1.0/500,
 	)
 
 	cs12 := constraint.SingleConstraint(&c12)
 	statsFunc123(
 		cs12,
-		"stats: [rows=20000000, distinct(1)=1]",
+		"[rows=20000000, distinct(1)=1]",
 		1.0/500,
 	)
 
 	cs123 := constraint.SingleConstraint(&c123)
 	statsFunc123(
 		cs123,
-		"stats: [rows=400, distinct(1)=1, distinct(2)=1, distinct(3)=5]",
+		"[rows=400, distinct(1)=1, distinct(2)=1, distinct(3)=5]",
 		5.0/125000000,
 	)
 
 	cs32 := constraint.SingleConstraint(&c32)
 	statsFunc123(
 		cs32,
-		"stats: [rows=80000, distinct(2)=2, distinct(3)=1]",
+		"[rows=80000, distinct(2)=2, distinct(3)=1]",
 		2.0/250000,
 	)
 
 	cs := cs3.Intersect(&evalCtx, cs123)
 	statsFunc123(
 		cs,
-		"stats: [rows=80, distinct(1)=1, distinct(2)=1, distinct(3)=1]",
+		"[rows=80, distinct(1)=1, distinct(2)=1, distinct(3)=1]",
 		1.0/125000000,
 	)
 
 	cs = cs32.Intersect(&evalCtx, cs123)
 	statsFunc123(
 		cs,
-		"stats: [rows=80, distinct(1)=1, distinct(2)=1, distinct(3)=1]",
+		"[rows=80, distinct(1)=1, distinct(2)=1, distinct(3)=1]",
 		1.0/125000000,
 	)
 
 	cs45 := constraint.SingleSpanConstraint(&keyCtx45, &sp45)
 	statsFunc45(
 		cs45,
-		"stats: [rows=1000000000, distinct(4)=1]",
+		"[rows=1000000000, distinct(4)=1]",
 		1.0/10,
 	)
 }
@@ -215,12 +212,7 @@ func testStats(
 ) {
 	t.Helper()
 
-	ev := ExprView{}
-
-	tp := treeprinter.New()
-	ev.formatStats(tp, sb.s)
-	actual := strings.TrimSpace(tp.String())
-
+	actual := sb.s.String()
 	if actual != expectedStats {
 		t.Fatalf("\nexpected: %s\nactual  : %s", expectedStats, actual)
 	}
